Consolidate new logic in redirect module
